requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LLM APIs
openai==1.3.5
anthropic==0.7.7

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2

# Vector Database & Search
qdrant-client==1.7.0
# weaviate-client==3.25.3  # Alternative vector DB

# Database
# pymongo==4.6.0  # For MongoDB
psycopg2-binary==2.9.9  # For PostgreSQL
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Drive Integration
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.110.0

# Embeddings
sentence-transformers>=5.1.0

# Utilities
tenacity==8.2.3
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
